Remove primary-only restriction from email_confirmed signal and migration Fix migration to actually set private=False on existing individual orgs Move charge_created signal handler from users to organizations app Inline perform_create in InvitationViewSet
