+++ squarelet/users/signals.py
@@ -1,6 +1,5 @@
 # Django
 from django.db import transaction
-from django.db.models.signals import post_save
 from django.dispatch import receiver
 from django.utils.translation import gettext_lazy as _
 
@@ -13,7 +12,6 @@ from hijack.signals import hijack_ended, hijack_started
 from squarelet.core.mail import send_mail
 from squarelet.core.utils import new_action
 from squarelet.oidc.middleware import send_cache_invalidations
-from squarelet.organizations.models.payment import Charge
 from squarelet.users.models import User
 
 registry.register(User)
@@ -50,12 +48,11 @@ def user_signed_up(request, **kwargs):
 
 
 def email_confirmed(request, email_address, **kwargs):
-    if email_address.primary:
-        send_cache_invalidations("user", email_address.user.uuid)
-        org = email_address.user.individual_organization
-        if org and org.pk and org.hidden:
-            org.hidden = False
-            org.save(update_fields=["hidden"])
+    send_cache_invalidations("user", email_address.user.uuid)
+    org = email_address.user.individual_organization
+    if org and org.pk and org.hidden:
+        org.hidden = False
+        org.save(update_fields=["hidden"])
 
 
 def email_changed(request, user, from_email_address, to_email_address, **kwargs):
@@ -85,15 +82,6 @@ def email_changed(request, user, from_email_address, to_email_address, **kwargs)
     )
 
 
-def charge_created(
-    sender, instance, created, **kwargs
-):  # pylint: disable=unused-argument
-    """Un-hide individual orgs when a charge is created"""
-    if created and instance.organization.individual and instance.organization.hidden:
-        instance.organization.hidden = False
-        instance.organization.save(update_fields=["hidden"])
-
-
 signals.email_confirmed.connect(
     email_confirmed, dispatch_uid="squarelet.users.signals.email_confirmed"
 )
@@ -105,9 +93,4 @@ signals.user_logged_in.connect(
 )
 signals.user_signed_up.connect(
     user_signed_up, dispatch_uid="squarelet.users.signals.user_signed_up"
-)
-post_save.connect(
-    charge_created,
-    sender=Charge,
-    dispatch_uid="squarelet.users.signals.charge_created",
 )
