+++ squarelet/users/tests/test_signals.py
@@ -6,29 +6,25 @@ import pytest
 from allauth.account.models import EmailAddress
 
 # Squarelet
-from squarelet.organizations.models.payment import Charge
-from squarelet.organizations.tests.factories import OrganizationFactory
 from squarelet.users import signals
 
 
 def test_email_confirmed(user_factory, mocker):
     mocked = mocker.patch("squarelet.users.signals.send_cache_invalidations")
     user = user_factory.build()
-    email = EmailAddress(user=user, primary=True)
+    email = EmailAddress(user=user, primary=False)
     signals.email_confirmed(None, email)
     mocked.assert_called_with("user", user.uuid)
 
 
 @pytest.mark.django_db
 def test_email_confirmed_unhides_user(user_factory, mocker):
-    """Confirming a primary email should set hidden=False on the individual org"""
+    """Confirming any email should set hidden=False on the individual org"""
     mocker.patch("squarelet.users.signals.send_cache_invalidations")
     user = user_factory()
     assert user.individual_organization.hidden is True
 
     email = EmailAddress.objects.get(user=user)
-    email.primary = True
-    email.save()
 
     signals.email_confirmed(None, email)
 
@@ -37,10 +33,12 @@ def test_email_confirmed_unhides_user(user_factory, mocker):
 
 
 @pytest.mark.django_db
-def test_email_confirmed_nonprimary_does_not_unhide(user_factory, mocker):
-    """Confirming a non-primary email should not change hidden status"""
+def test_email_confirmed_nonprimary_unhides_user(user_factory, mocker):
+    """Confirming a non-primary email should also unhide the user"""
     mocker.patch("squarelet.users.signals.send_cache_invalidations")
     user = user_factory()
+    assert user.individual_organization.hidden is True
+
     email = EmailAddress.objects.create(
         user=user, email="secondary@example.com", primary=False, verified=True
     )
@@ -48,35 +46,7 @@ def test_email_confirmed_nonprimary_does_not_unhide(user_factory, mocker):
     signals.email_confirmed(None, email)
 
     user.individual_organization.refresh_from_db()
-    assert user.individual_organization.hidden is True
-
-
-@pytest.mark.django_db
-def test_charge_unhides_individual_org(user_factory):
-    """Creating a charge for an individual org should set hidden=False"""
-    user = user_factory()
-    org = user.individual_organization
-    assert org.hidden is True
-
-    signals.charge_created(
-        sender=Charge, instance=Charge(organization=org), created=True
-    )
-
-    org.refresh_from_db()
-    assert org.hidden is False
-
-
-@pytest.mark.django_db
-def test_charge_does_not_unhide_group_org():
-    """Creating a charge for a non-individual org should not change hidden"""
-    org = OrganizationFactory(individual=False, hidden=True)
-
-    signals.charge_created(
-        sender=Charge, instance=Charge(organization=org), created=True
-    )
-
-    org.refresh_from_db()
-    assert org.hidden is True
+    assert user.individual_organization.hidden is False
 
 
 @pytest.mark.django_db(transaction=True)
